pyproject.toml
[project]
name = "news-scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "edge-tts>=7.2.7",
    "google-genai>=1.59.0",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "pygame>=2.6.1",
    "typing>=3.10.0.0",
]
